[#14] Feat: MockupNetworkManager 대체
